Add create route
With the new route, now is is possible to create new entries for monkeys. Since there is not validation of request yet, it is necessary to pass at least id and species parameters, so that DynamoDB can save the new Item.
There was also changes on serverless.yml so that the LambdaFunction can insert/read/delete items from DynamoDB.